tool: move attr parsing to `mod attr`
